move ModelsService to modelsService.ts
