+++ lib/shared/src/llm-providers/utils.ts
@@ -1,5 +1,5 @@
 import type { CompletionsModelConfig } from '.'
-import { modelsService } from '../models'
+import { modelsService } from '../models/modelsService'
 
 export function getCompletionsModelConfig(modelID: string): CompletionsModelConfig | undefined {
     const provider = modelsService.getModelByID(modelID)
