+++ lib/shared/src/models/utils.ts
@@ -1,6 +1,6 @@
-import type { ModelRef, ModelRefStr } from '.'
 import { ModelTag } from '..'
 import type { Model } from './model'
+import type { ModelRef, ModelRefStr } from './modelsService'
 
 export function getProviderName(name: string): string {
     const providerName = name.toLowerCase()
